pyproject.toml
[project]
name = "autoresearch-council-arena"
version = "0.1.0"
description = "Autonomous writing optimizer: autoresearch experiment loop + llm-council evaluation pipeline"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
